# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShopIA API",
    "version": "1.0.0",
    "description": "Documentation API ShopIA Dashboard"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/api/boutiques/shopbot/campagnes": {
      "get": {
        "summary": "Campagnes WhatsApp actives",
        "description": "Liste les campagnes WhatsApp en cours et planifiées",
        "tags": [
          "ShopBot"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "boutique_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "active",
                "planifiee",
                "terminee"
              ],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campagnes récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "nom": {
                            "type": "string"
                          },
                          "boutique": {
                            "type": "string"
                          },
                          "statut": {
                            "type": "string",
                            "enum": [
                              "Active",
                              "Planifiée",
                              "Pause",
                              "Terminée"
                            ]
                          },
                          "date_debut": {
                            "type": "string",
                            "format": "date"
                          },
                          "date_fin": {
                            "type": "string",
                            "format": "date"
                          },
                          "destinataires": {
                            "type": "integer"
                          },
                          "taux_ouverture": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/boutiques/prestataires/stats": {
      "get": {
        "summary": "Statistiques globales des prestataires de service",
        "description": "Retourne les KPIs agrégés pour le suivi des prestataires de service.\n\n**Métriques** :\n- Prestataires actifs\n- RDV actifs\n- Réservations\n- Taux d'occupation\n",
        "tags": [
          "Prestataires"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "prestataires_actifs": {
                          "type": "integer",
                          "example": 10
                        },
                        "prestataires_actifs_tendance": {
                          "type": "number",
                          "example": 3
                        },
                        "rdv_actifs": {
                          "type": "integer",
                          "example": 50
                        },
                        "rdv_actifs_tendance": {
                          "type": "number",
                          "example": 5
                        },
                        "reservations": {
                          "type": "integer",
                          "example": 30
                        },
                        "reservations_tendance": {
                          "type": "number",
                          "example": 8
                        },
                        "taux_occupation": {
                          "type": "number",
                          "example": 76.5
                        },
                        "taux_occupation_tendance": {
                          "type": "number",
                          "example": 3
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}